provision/docker: remove UploadDeployer, ArchiveDeployer and RebuildDeployer
